requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2
requests
python-dotenv
pytest
instagrapi==2.2.1
Pillow
itsdangerous
tenacity